requirements.txt
anthropic
chromadb
sentence-transformers
gradio
python-dotenv
requests
feedparser
PyPDF2
